added the "-root" option. See the following feedback -------------------------- Date: Wed, 24 Jul 2002 09:38:58 +0200 From: Marc Fleischeuers <M.Fleischeuers@devote.nl> Subject: feature request for Sun XML Generator To: kohsuke.kawaguchi@eng.sun.com Cc: r.bosman@devote.nl
Dear Mr. Kohsuke,

Thanks a lot for your XML generator, it looks like it will save us quite a
lot of time and effort. There is one feature however that we require to use
the program fully: we would like to be able to specify with which top-level
element the generator starts generating. Let me clarify, by explaining our
schema.

We have a schema that contains some 30 top-level elements. In our input, we
know that only one of these elements is actually present as top-level
element. The other elements serve as targets to be ref'd from within this
one 'true' top-level element. For parsing, this works just fine; generating
however is a problem: there is only 3% chance that the intended element is
chosen to start with.

Is it possible to extend your program, so that we can specify the initial
branch in the schema? That would mean a lot to us, an (I can imagine) other
users as well.

Our next step would be to programmatically call the generator, and use is as
input for our SAX contenthandler. Is there any documentation on how to do
that?

I hope you can help us. If you reply, can you please 'reply to all' so my
colleague Rob Bosman is also notified?

TIA -
Marc Fleischeuers
Ordina Finance Business Solutions - E-business and IT Architects


git-svn-id: https://svn.dev.java.net/svn/msv/trunk@1433 8cecbe3e-ef64-0410-a610-ece5afc49ac1